+++ src/i18n/routing.ts
@@ -1,5 +1,4 @@
 import { defineRouting } from 'next-intl/routing'
-import { createNavigation } from 'next-intl/navigation'
 
 export const routing = defineRouting({
   locales: ['en', 'ru'],
@@ -22,6 +21,3 @@ export const routing = defineRouting({
     },
   },
 })
-
-export const { Link, redirect, usePathname, useRouter } =
-  createNavigation(routing)